5382: move files
